requirements.txt
Django==5.2.6
djangorestframework==3.15.2
scikit-learn==1.5.2
pandas==2.2.3
nltk==3.9.1
pytest==8.4.2
pytest-django==4.11.1
requests==2.32.3
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.7.0